ft: add provisioning dispatcher class
This class is used to provision resources (given by arbitrary names)
to a set of subscribers, in a fair way. The primary purpose now is to
allocate raft sessions to running queue populator processes, and
reallocate automatically to others when some of them fail to keep
processing all raft sessions.

The implementation is based on zookeeper.