ft: ProvisionDispatcher class in queue populator
Use the ProvisionDispatcher class to configure the queue populator so
that it manages a raft session ID. Make it dynamic so that a running
queue populator can switch to another raft session ID any time (if the
dispatcher reassigns a new one, after a failure for example), or can
become idle as well.

The queue populator also becomes idle if there is more than one raft
ID in the managed list.

Add support for graceful handling of SIGTERM, to unsubscribe from the
dispatcher so that the dispatcher can quickly notify other queue
populators to take over the unmanaged raft ID.