ft: ProvisionDispatcher class in queue populator
Use the ProvisionDispatcher class to configure the queue populator so
that it manages one or more raft session IDs. Make it dynamic so that
a running queue populator can switch to another set of raft session
IDs any time (if the dispatcher reassigns a new one, after a failure
for example), or can become idle as well if no raft ID is assigned
from zookeeper.

Add support for graceful handling of SIGTERM, to unsubscribe from the
dispatcher so that the dispatcher can quickly notify other queue
populators to take over the unmanaged raft IDs.

- refresh behavior test

Behavior tests were broken, fix them regarding latest changes on queue
populator.