ft: enable keep-alive mode for backbeat AWS client
This will allow to re-use a pool of connections both for the source
requests and the destination requests.

To do this we simply setup two global HTTP agents, one for the source
and one for the destination, and attach them when creating the AWS
client object.

For SSL support we will need an HTTPS agent instead.

Tested manually by spawning an ad-hoc server with 'nc' on the
destination.