ft: use a bootstrap list for source/target S3 and Vault
For now, only the first element of the list is used, this will allow
to update Federation to provide a list of hosts in the meantime, later
we will refine to actually use multiple hosts from the list.

For source it does not matter but some of the code shares processing
of host/port from conf, so it makes more sense to have both source and
destination use the same convention in config.