ft: auto-create zookeeper base paths for zenko
Auto-create queue populator base path on zookeeper if it does not
exist, when starting up the queue populator deployed previously with
docker swarm for zenko.

Since docker and docker swarm do not provide an easy way to run
specific commands at deployment time, especially conditioned to
containers being up and running first, in order to ensure zookeeper
base paths are created initially, we wrap the zookeeper client handle
creation with some custom code to do this.

This is enabled only in zenko mode with docker swarm deployment,
default options (including Federation-based deployment) do not change
the existing behavior which is to not create the base path if it does
not exist (and then further zookeeper commands shall fail with NO_NODE
errors).