ft: implement "echo" replication mode
!POC only!

This mode processes the "create bucket" entries on the source to
configures CRR automatically for this bucket through backbeat
(including creating the destination account if it does not exist, then
basically what does the replication.js setup script).

The destination bucket has the same name than on the source, and the
owner account has the same display name and email address (but a
different user ID and canonical ID since on a different Vault).