ft: support for get and put metadata blob
With the introduction of one to many crr, there will be a status update for each backend.
Since there will be multiple requests to the source to update the replication status, it's important
to get the latest copy of the metadata blob as previous updates would render the copy it's holding stale.