bf: remove broken design with unlockBucket messages
The issue with the current design is that we only send one
'unlockBucket' message after the bucket has been listed entirely, but
nothing guarantees that previous listing tasks executed by other
producers will not be longer to proceed, effectively voiding the
purpose of knowing when all objects of this bucket have been
processed.

One case where this is very likely to happen is when a specific range
of objects in the bucket have to be expired at once (other than the
last listed range), because it will take longer to process objects for
that particular range than it will be to finish listing the bucket and
processing the expirations for last range. Moreover this is a common
case in production (e.g. expiring a sub-directory).

An alternative implementation based on monitoring the backlog of kafka
to apply backpressure will be done instead.