bf: switch to node-rdkafka
kafka-node has issues dealing with reconnections etc. especially when
there are concurrent joins of multiple consumers of the same
group. This often leads to multiple consumers processing the same
partition, which is not just suboptimal, but breaks lifecycle
processing.

node-rdkafka is a more maintained and better tested library, relying
on librdkafka underneath.