bf: backpressure for lifecycle
In order to avoid accumulating lifecycle tasks and diverging when the
system cannot process everything in the allocated time, add a
mechanism to skip processing cycles when the previous cycle has not
finished yet.

Kafka consumers in LifecycleConsumer and LifecycleProducer now publish
to zookeeper, for their consumed topic partitions, their current
committed offset and the latest topic partition offset.

The conductor checks those when he's about to start a new cycle: if
they are not equal (i.e. if the backlog is non-null), it means there
are still things being processed, so the conductor does not trigger a
new cycle and waits for the next one.