fix: temp fix for listMultipartUpload
Bug found in listMultipartUpload. In the interest of time,
I've added a temporary fix in this PR but this will need
to be addressed in S3/Arsenal/Metadata depending on how
the listMultipartUpload is implemented.

Normally when listing uploads and specifying markers, the
listing should begin the listing with not the specified
marker, but the upload right after it. This is the
behavior I saw testing with aws directly, and compared
with the behavior I saw with our S3 implementation.