bf: ZENKO-316 use zk dispatcher for mongo log
We need only and exactly one queue populator processing the mongo log,
but we have 10 running in Federation deployments, so one has to be
picked and failover to another when the active one fails.

This can be implemented the same way we did for raft session log
consumption, with help from zookeeper watchers (ProvisionDispatcher),
but instead of having 8 raft sessions to consume from, we only have
one mongo log.

Kubernetes deployments don't need this, so an option
"subscribeToLogSourceDispatcher" has been added, set to true in
Federation deployments only.