ft: ZENKO-143 force read on local source
Make use of x-amz-location-constraint header to force a read from the
local source, in case a cloud target is the preferred read location
e.g. w/ transient source.

- If the cloud target replication is PENDING it avoid failing the read

- If it's in COMPLETED status it enforces a local read for transient
source to be effective.

For this to work, we have to go through a custom backbeat route that
maps to AWS SDK getObject() with support for non-standard
x-amz-location-constraint header.