bf: S3C-1550 race condition with raft session provisioning
Fix a case of race condition in the zookeeper provisioning code
responsible to assign raft session IDs to the queue populator
instances.

The redispatch of new provisions to their respective owners is done
asynchronously, this can cause conflicts with new owners showing up
during this process. To fix, prevent from running two concurrent
redispatch, wait for the first to complete if a new redispatch is
needed.

Re-enable provisioning test on the CI