bf: ZENKO-860 replace replication batching by continuous task flow
Keep consuming kafka messages as soon as new workers become available
in the queue, rather than relying on the 'drain' callback from
async.queue.

This prevents being stuck from consuming new messages in case one of
the tasks takes a long time to finish (e.g. if retrying on a failing
cloud).

Change the way we manage Kafka consumer offsets: instead of disabling
auto-commit by default, always enable it and disable automatic offset
store of consumed messages. This allows to set the offset based on the
processed messages rather than the consumed messages, relying on a
helper class OffsetLedger to maintain knowledge of the highest offset
that can be committed, i.e. that is a higher-bound of already
processed message offsets.