bf: ZENKO-1005 send noop entries for new locations
On new consumer group startup, on init of consumers, send
canary/noop entries to set initial offset for the new
group.

The method used here is similar to how bootstrap entries
are used in tests. One main difference is on startup,
sending the canary messages once making sure the
consumers have been assigned or have started to be
assigned to partitions.