ft: ZENKO-1056 return pause/resume status from api
When a user makes a pause/resume request to the backbeat
api, the api side will now make checks in zookeeper in
order to determine if the redis publish message needs to be
made. If a status change is expected, a retry status check
(up to 5 times of 1 second intervals) will be made in order
to make sure the status has changed as expected and will
return to the user the new status change.

Previously, if a user made a post request to change status,
the status change was not returned to the user. The user was
required to make a separate get request to check that status
did indeed change.

If status does not change within the retry period, an error
showing which sites did not update "in time" will be
returned.