ft: ZENKO-860 improve replication task pool efficiency
By reducing drastically the time it takes for a replication task to
complete, for the same concurrency number, a lot more tasks can be
executed, leading to more throughput and quicker completion of
replication queue tasks especially for small objects, while using
cloud resources more consistently and efficiently with the same bound
on the number of open connections.

To achieve this reduction in execution time, tasks do not wait anymore
for a delivery report to be returned from Kafka producers (for metrics
and backbeat-replication-status topics) before finishing. Most of the
wait time was spent waiting for the delivery reports, which are only
issued once Kafka commits changes in batch (every few seconds), hence
the tasks in progress were mostly waiting for those delivery reports
instead of transferring bytes of data between locations.

Instead, after the data replication is completed (success or failure),
and the task completes, we keep track of the delivery report pending
for the message produced to backbeat-replication-status topic. Only
when this status comes back (success or failure), we acknowledge that
this message has been consumed and processed, to allow eventually
committing the consumer group offset forward.

Delivery reports for messages produced to metrics topic are ignored,
because it does not seem to bring any consistency benefit to wait for
them (but we could instead implement a dedupe cache to get rid of
duplicates).