bf: S3C-1886 fix "processing" metrics retention
The fix consists of a refactor of how metrics are published, in a way
that maps what has been done in 8.0 branch but has not been backported
to 7.4, however the differences between the two branches called for a
custom implementation on 7.4, a simple cherry-pick was not possible.

A simpler fix removing a conditional check on publishing 'metrics'
event would have been possible, but it would not have been easily
testable and would have kept the technical debt that has been fixed in
8.0 branch (i.e. that BackbeatConsumer should not deal with
CRR-specific semantics).

Previously the metrics were published as a "metrics" event in the
BackbeatConsumer class, possibly gathering multiple stats together
before publishing. The condition to publish them was on
"messagesConsumed > 0", which was incorrect as new metrics could have
been published after "messagesConsumed" was reset to 0. By moving the
metrics publishing outside of BackbeatConsumer and in the CRR
extension code, this gets fixed on the way, and it becomes possible to
test published metrics without relying on a specific topic name being
published to.