refactor: ZENKO-734 remove destEntry from MultipleBackendTask
The "destEntry" param is a relic from ReplicateObject base class, that
does not play a role in zenko replication code: "destEntry" in CRR
code represents the metadata pushed to the target site, which has no
materialization in zenko replication as every operation is carried out
by the multiple backend layer.

Also, when it comes to changes related to using the queue processor
for lifecycle transitions, there are more reasons why it can become an
important cleanup.

This commit is a pure refactor, the few extra functions directly come
from ReplicateObject class without their "destEntry" and CRR-specific
behavior, that I deemed appropriate to "duplicate" to avoid
possibilities of code path incompatibility between both classes and
separate the matters better between zenko replication and CRR.