ft: ZENKO-1420 ActionQueueEntry model for background jobs
Provide a helper class to represent an action message that can be
passed along to backbeat services through Kafka queues to execute
specific actions. It can also represent the outcome of the actions
(status and potential results) returned by the service in another
queue.

It's meant to be generic to any kind of background job executed
through a service, but provides a structure that should eventually
help with instrumentation (tracing notably) and logging, while
improving consistency of our internal message passing.

The first user will be the "data mover" as a service called for
transition policies, where we'll be able to request a data copy to a
new location, and get back a result message with the access details to
the new data location, that the transition policies code can use to
update object metadata afterwards.