cleanup: ZENKO-1420 more generic BackbeatMetadataProxy.putMetadata()
Make the putMetadata() call more generic by taking separate params
like bucket name etc. and a metadata blob, instead of using an
ObjectQueueEntry directly. This will be useful to update metadata
based on action message fields that will not be converted into an
ObjectQueueEntry.

For consistency, also renamed "encodedVersionId" to "versionId" in
getMetadata(), as the encoded form is the common form used in AWS API
the name can be kept short.