refactor: ZENKO-1420 only one producer in bucket processor
Since we're going to publish to a third topic and we now have a way to
use a single producer to publish to multiple topics, convert the code
to only use one producer instance. Rework a bit some helpers to send
messages so that they are defined in the LifecycleTask class directly.