feature: ZENKO-1420 data mover service
Create a new task type inside the queue processor to copy data between
locations.

It relies on ActionQueueEntry class to marshall/unmarshall requests
and responses to/from the backbeat-data-mover topic.

The current version does not allow choosing the key of the new object,
as it is chosen by the multiple backend code that executes the copy,
but it fits the need for transition policies.