feature: ZENKO-734 lifecycle object processor for transitions
Update the lifecycle object processor to support transition policies.

The object processor is a client of the data mover: when a data copy
is completed, it receives the results in the
lifecycle-object-tasks-topic, and updates the location in metadata
according to the new location sent by the data mover, on success, and
garbage-collects the old location.