bugfix: ZENKO-1522 fix CRR on copied MPU object
Change the logic to ignore the 'MPU' flag set in
replicationInfo.content array (it may not always be set), instead rely
on the object size and if a MPU or not based on its content-md5
metadata attribute containing a dash.

Note: the copy code lacks integrity checking now, this would need a
rework but it's out of scope of this fix.