rf: ZENKO-1531 mongo processor use public methods
Previously fetched owner info only from the BucketInfo
instance and rather than using the public methods available
on the class instance, I updated the private variables
instead. This is a refactor to properly use the
public methods available for setting values.

Also, pass BucketInfo instance to the next helper methods
(i.e. `_processObjectQueueEntry`) so that we have access
to all the information on the Zenko Bucket. We will need
this for updating fields like acls or replication info.