bf: ZENKO-1629 ingestion repInfo md only updates
Identify md-only (i.e. PUT/DELETE tagging) updates for
setting the ReplicationInfo "Content" field. This is
necessary for replication-enabled ingestion buckets
in order to replication md-only updates.

Idea is to fetch metadata from mongo and make comparisons
between the new entry and existing entry in mongo. If entry
does not exist in mongo, this is a newly ingested entry.
If entry exists in mongo, we only compare tagging between
the two entries as that is the only field we care about.