bugfix: ZENKO-1582 skip update if new location == old location
It's possible to get duplicate updates, not always because of kafka
delivering duplicate messages, but also if a duplicate copy occurred
(possible even if we reduced the risk by deduplicating in the queue
processor).

In such case we need to skip the GC of the location otherwise we lose
the valid data.