ft: ZENKO-1448 cleanup ingestion state on removal
When an ingestion bucket is removed from Zenko, we should
cleanup zookeeper state in case of scenarios where maybe
later on, another ingestion bucket is created with the
same name as previous bucket.

We do not cleanup pause/resume state because this is
set by location storage. Potentially, a user could create
multiple ingestion buckets for the same storage location.