bugfix: ZENKO-1601 fix replication status processor skipping entries
Modify the ReplicationTaskScheduler class with custom "queue key" and
"dedupe key" user functions: that allows the flexibility for queueing
and/or skipping tasks based on the provided keys. Also take the
opportunity to move it to global backbeat lib and rename the class to
TaskScheduler, because it's not specific to replication anymore and
might be potentially useful for other purposes.

The queue processor now passes explicitly queue key and dedupe key
functions, while the replication status processor only passes a queue
key (because it does not want deduplication).

I reworked the unit tests by splitting them conceptually to test the
scheduler and the key-generating functions separately. This makes them
simpler and more focused on testing each component.

On the scheduler tests, I think by reading the code that the removed
tests (recently introduced) were not effectively testing the queueing
behavior anymore, so I though cleaner to revert back to the previous
tests and add an extra test for deduplication.