bugfix: ZENKO-1684 do not wait for GC message delivery
In replication status update tasks with transient source (8.0+) and in
lifecycle transition update (8.1+ only), do not wait until the GC
acknowledges the delivery of the messages to return a status.

There's no need to wait and it slows down considerably the processing,
as delivery takes about 2 seconds to acknowledge a message, so with 10
workers it limits processing to 5 messages per second.