feature: ZENKO-1520 expose data mover metrics
Add queued and processed metrics to the data mover.

In order to have the data mover completely manage its own metrics and
have a cleaner API, created a ReplicationAPI class containing methods
to create action entries and to send them to the data mover queue,
with a list of documented parameters.