bf: ZENKO-1744 ingestion kafka producer polls
rdkafka has an in-mem queue or buffer that will store stats.
Poll will read from this queue or buffer and remove. With
ingestion, if we have a high number of entries being read,
this queue fills up fast.

To help deal with this, we will lower the poll interval ms
to poll more frequently.