bugfix: ZENKO-1907 fix kafka backlog check
Fix kafka backlog check with removed locations or other situations
where nodes become unused (like with partitions waiting to be
consumed).

The fix consists of ignoring (and deleting) nodes that have been
updated since longer than twice the interval of the offset publishing
task. Since those updates normally occur regularly, nodes that did not
get updated are normally obsolete and should be ignored and deleted.

Also removed the check that forces the first batch to be executed
regardless of the backlog, it's now unnecessary since obsolete nodes
will be ignored, and it's better to enforce the check after a restart.