improvement: S3C-2837 gather and log kafka consumer lag metrics
With an option to the BackbeatConsumer instance, periodically log the
total consumer lag, as the sum of the individual partitions lag as
returned by librdkafka in its statistics event.

It's not clear what is the behavior with multiple consumer groups, or
if the consumer instance retrieving the statistics has an influence on
the results. It looks to be returning the lag correctly for a single
consumer group and that is the case for the backbeat-replication topic
in S3C.