bugfix: S3C-2775 provide versionID to 'PUT /_/backbeat/metadata' request
Add the versionId in the query string on 'PUT /_/backbeat/metadata'
requests.

The goal is to be correct regarding the check done in cloudserver
backbeat routes: instead of checking the master version metadata, it
will check the target version object, resulting in a ObjNotFound error
if the specific version does not exist. This in turn allows to fix
UTAPI counting of replicated objects, by providing the information of
whether an existing object with the same version exists.

(cherry picked from commit c33e26ead0cad15c7e4aaa66181270b7dc3fbe59)