bugfix: S3C-3685 stop reading new log entries after provisioning update
Do not trigger a new batch in case a provisioning update is pending:
instead, stop after the current batch to give a chance to update the
provisioning and consume from the new (set of) raft sessions that the
readers are responsible for.

Also improve logging by outputting a status log line after every
batch, previously the log only occurred after a series of batches that
completes reading the whole log and that were publishing at least one
entry to a kafka queue.