bugfix: S3C-2665 helper for removing orphans cleanly
Helper function that will be needed to remove all orphaned data after
a replication failure. The async.mapLimit function does not wait for
pending requests after an error, which makes it not suitable since we
need to wait for pending PUT rquests that also will need to be cleaned
up.