test: S3C-4425 test for failed CRR mechanism
Check that a correct message is sent to the
`backbeat-replication-failed` queue when a failure to replicate occurs
and a status FAILED is to be written, by extending a queue processor
functional test.

This adds some coverage for the FailedCRR mechanism, it does not
specifically test the fix in S3C-4425 but adds regression testing.