bugfix: S3C-4493 RoleCredentials: fix error handling
Fix by returning a copy of InternalError from arsenal errors.

Make sure we return a copy of the arsenal error to avoid side effects
with AWS SDK modifying the error properties.

Also fix the unit test to actually trigger a non-arsenal error like
described, as previously the "socket.destroy()" was caught as an
arsenal error in Vaultclient.