bugfix: BB-70 upgrade node-rdkafka to 2.10.1
More recent versions (2.9.0+) have fixed leaks that affected the queue
populator and required regular restarts to avoid OOM conditions.

The latest version to date is 2.11.0, however it is incompatible with
our current nodesvc-base image due to the GCC version not supporting
the flag "-std=c++14", hence upgrading to 2.10.1 until we have an
upgraded nodesvc-base image.