bugfix: retry S3 calls from lifecycle processors
Only applies to retryable errors though, other errors will potentially
interrupt the scan, which will resume on the next lifecycle batch
trigger. This is on purpose, as fatal errors being retried over and over
again have the potential to increase the load to the point of causing
cascading failures, while not performing actual work.