improvement: BB-122 pass "client.rack" property to kafka consumer
To enable follower fetching on stretched configuration, pass the
"client.rack" property to kafka consumers from the "site" attribute
configured from Federation.

There is no unit test attached as this is barely passing a
configuration param to Kafka consumer client and would not check if
the param was applied all the way. The envisioned way to test this
properly (almost) end-to-end would be once Integration runs a 2-site
configuration to check if the backbeat and bucket notification logs
indeed contain the "follower fetching enabled" log line, asserting
that the configuration was applied for the relevant topics when on a
2-site config.

It does not look like an option either to check if the consumer is
always fetching from a local replica, because Kafka does not
systematically consume from a local replica but only when it sees fit
based on other considerations (load balancing etc.).

Note: to make it work, it also requires changes on Federation side to
set the replica selector class and to add the "site" property to
backbeat configuration.