feature: BB-167 add dynamic kafka-connect connector configuration
Changed the oplog populator to dynamically configure the kafka-connect mongo
connector to listen to specific buckets that use at least one of the active
backbeat extensions.

All buckets that use at least one extension are first listed and then added to
the configuration of the kafka-connect mongo connector (only one connector is
used for now).

A change stream is then established to listen to changes on the __metastore
collection. The connector is then dynamically reconfigured to start or stop
listening to buckets that have been added, removed or updated in the collection.