feature: BB-164 add OplogPopulator ConnectorsManager class
A ConnectorsManager keeps a list of active connectors

It is also used to spawn and configure connectors according to config.
Old connectors are first retreived and reconfigured, then if more
connectors are required it creates them.

Old connectors are reconfigured where buckets that are no longer valid
are removed from the connectors' config