bugfix: BB-187 BN: add object key to keyed partitioning
When pushing to the internal backbeat-bucket-notification topic, use
"bucket/key" string instead of just "bucket" for keyed
partitioning. This allows multiple objects belonging to the same
bucket to go to different notification partitions. It retains the
order of operations on a single object, or multiple versions of that
object, by pushing those operations to the same partition.

This complements the previous commit that fixed keyed partitioning for
the Kafka destination.

(cherry picked from commit 6c161840e7b52872a17024edb9a021e4a305bb95)