feature: BB-213 fix kafka-connect connectors not being configured correctly
A race condition occurs when multiple events are processed at the same time
and that affect the same connector, in which case the connector is not configured
correctly.

For example when two buckets are being added to the same connector at the same time,
only one of them gets added, this is caused by the two operations getting and updating
the list of buckets assigned to the connector at the same time, which results in one of
them overwriting the other.

The solution is using a scheduler that periodically updates all of the connectors.
Connectors are no longer updated on an event basis