Run all ReplayProcessors in the same process
Change the deployment model for replay processors, from 1 process per
target (S3C/7.x) to 1 process for all targets (Zenko/8.x), like it is
done for the queue processor.

Since the task.js entrypoint is the same, deduplicate the code to keep
only the one from QueueProcessor, by taking an optional param on command
line to specify the "replayTopic" to process: if not specified, it will
simply process the replication topic (as before), otherwise it will be a
replay processor.

Issue: BB-234