Make ECONNRESET errors retryable
The ECONNRESET (and associated EPIPE and ETIMEDOUT) errors come from the
http agent, and are not automatically tagged retryable with AWS-SDK [1].

This is especially problematic since the error may come from unlucky
timing on the keep-alive socket [2] (e.g. socket getting closed while it
is about to be reused) or an (over)loaded remote service: and we should
thus try again.

[1] https://github.com/aws/aws-sdk-js-v3/issues/1196#issuecomment-636589545
[2] https://nodejs.org/api/http.html#requestreusedsocket

Issue: BB-292