bugfix: BB-159 update notificationQueuePopulator to use the new update event
The update operation we do before deleting objects will be used as a deletion
event. This helps determine wether an object is versioned, non versioned or
version suspended.

To avoid pushing two notifications for the same events we ignore all versioned
masters and only process the versions. The only cases where master events are
processed is when the master is non versioned or when it's a null version