Discard restore expiry request if date not reached
There is a sort of race condition if the `s3:restore` api is used to
update the expiration date of the restore object at the last moment,
just while it is in-flight: e.g. after it is queued for expiration, but
before this queue message gets processed by
backbeat.

To avoid this, we do an extra check and simply discard the message in
that case, relying on cold storage to queue the object for
expiration again once it will have reached the "new" expiration date.

Issue: BB-402