keep extra fields added manually in kafka connector config
When connectors stop consuming the oplog because of the outdated
resume token issue, we can only unblock them by renaming the partition
used to store the offsets. This is done by adding an extra parameter in
the connectors' config.

This parameter is not included in the config by default, and we want to
keep it from changing when the oplog populator restarts.

Issue: BB-416