bf: BB-419 notification: increase default BackbeatConsumer concurrency
Notification QueueProcessor was limited by a concurrency of 10, which
meant it could only wait for 10 delivery reports from the producer
before consuming more entries.

Increased the default concurrency to 1000 to improve throughput and
add a test to show that it behaves well with this number of pending
reports (the processing does not involve anything else asynchronous
than waiting for delivery reports), as well as a few other extra
tests.

Also fix and improve logging related to processing errors and error
reporting from delivery reports.