BB-419 [impr] configurable notif Kafka poll interval
Make the external Kafka producer poll interval configurable: while the
default is 2000ms, it could be helpful to reduce its value, e.g. to
reduce latency of delivery in case the external Kafka cluster cannot
receive all notifications in time.

It's not clear at this point what would be a better value due to lack
of testing to the limits, and 2000ms doesn't look unreasonable as a
trade-off between latency and polling overhead, so leaving the default
value as it is. It is however important to have a concurrency value
high enough to not limit the possible throughput per notification
processor, which is limited by "concurrency / poll_interval_seconds"
messages per second. The concurrency has been raised to 1000 by
default which therefore allows 500 messages per second to go through
per processor with a 2000ms poll interval.